Added a line to the README for importing the project.
Added a gitignore to not commit a directory called drawn.  This will be used to store any images that are hand drawn, which I don't want commited to git.

Changed the Settings file to start using checkboxes (although I don't think this works right at the moment.)

Changed the Main menu file to seperate out a few things for clearity.